Refactor host functions and systems for improved clarity and consistency
- Removed unused imports in `host_functions.rs`.
- Updated `host_systems.rs` to use `SpriteConfigStrIds` instead of `SpriteConfigStringIds` and adjusted related logic.
- Changed state references from `NoGameSession` to `NoSession` in `main_menu.rs`, `main_menu_layout.rs`, and `lobby_systems.rs`.
- Refactored sprite systems to utilize a new `SpriteConfigStrIds` structure for better type safety and error handling.
- Introduced new material types in `tilemap` for enhanced texture handling, including `MonoRepeatTextureOverlayMat` and `TwoOverlaysExample`.
- Cleaned up terrain generation systems and removed deprecated materials.
- Updated Cargo.toml files to include new dependencies and removed unnecessary ones.
- Consolidated system chains for better performance and readability across various modules.

+++ crates/tilemap/src/tile/tile_resources.rs
@@ -1,4 +1,4 @@
-use bevy::platform::collections::HashMap;
+use bevy::{math::f32, platform::collections::HashMap};
 #[allow(unused_imports)] use bevy::prelude::*;
 #[allow(unused_imports)] use bevy_replicon::prelude::*;
 #[allow(unused_imports)] use bevy_asset_loader::prelude::*;
@@ -65,8 +65,8 @@ pub struct ShaderRepeatTexSerisHandles {
 pub struct ShaderRepeatTexSeri {
     pub id: String,
     pub img_path: String,
-    pub scale: u32,
-    pub mask_color: [u8; 4],
+    pub scale: f32,
+    pub mask_color: [f32; 4],
 }
